backend/pyproject.toml
[project]
name = "imgrep-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "colorlog>=6.9.0",
    "dotenv>=0.9.9",
    "faiss-cpu>=1.11.0.post1",
    "flask>=3.1.1",
    "flask-sqlalchemy>=3.1.1",
    "mypy>=1.16.1",
    "opencv-python>=4.12.0.88",
    "psycopg2-binary>=2.9.10",
    "rapidfuzz>=3.13.0",
    "scikit-image>=0.25.2",
    "torch>=2.7.1",
    "torchvision>=0.22.1",
    "werkzeug>=3.1.3",
    "spacy>=3.8.7",
    "python-dateutil>=2.9.0.post0",
    "nltk>=3.9.1",
    "facenet-pytorch>=2.5.3",
]
[tool.mypy] 
strict = true
